// src/types/generated/EffectiveScheduleResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplicationRule } from "./ApplicationRule";
import type { ScheduleLibraryItem } from "./ScheduleLibraryItem";

/**
 * Response with effective schedule for a date
 */
export type EffectiveScheduleResponse = { library_item: ScheduleLibraryItem, specificity: number, rule: ApplicationRule, };