// src/types/generated/CalendarDayScheduleMatches.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CalendarScheduleMatch } from "./CalendarScheduleMatch";

/**
 * All matching schedules for a calendar day
 */
export type CalendarDayScheduleMatches = { winning_match: CalendarScheduleMatch, other_matches: Array<CalendarScheduleMatch>, };